Update to loading data
I created a file that can be run that will do the data pre-processing and turn that file into a pickle file.  This is because the lemmatizing process take a very long time.  Once the pkl file is created, we can play around with the creation of algorithms and train from the already preprocessed data.