Adding updated data files
Just ensuring that we have the required files loaded for data.  The ones with the "clean" as part of the file name are data files that have proper names and titles redacted.  However, they currently have issues and aren't used until those formatting and data quality are cleaned up.  Pickle files are used so that constant pre-processing when running the models isn't required.