Naive Bayes model update
So, this update is predominantly about converting the .py solution to a jupyter notebook for ease of running.  The functionality that is a part of the python project has been refactored into the notebook.  Further, the notebook contains the pre-processing step that can be circumvented if you don't want to run that step by going to the point where it uses the pickle file.  It also has steps where it tries different hyperparamter settings and produces heatmaps to better understand what's happening in this baseline model.